Fix Geometry Bugs (#19)
* Initial fix of geometry that seems to work so far. Must test DT and nested assemblies.

* Adds new Cross surface to Python API

* Fixes bug with DT/CT

* Starts new BEAVRS model made with Python API

* Applies a clang-format

* Small formatting changes.
