+++ include/geometry/geo_lily_pad.hpp
@@ -34,12 +34,18 @@ struct GeoLilyPad {
   enum class PadType { Universe, Lattice, Cell };
 
   GeoLilyPad() {}
-  GeoLilyPad(PadType t, uint32_t i, Position r, std::array<int32_t, 3> ti,
-             bool ou)
-      : type(t), id(i), r_local(r), tile(ti), in_lattice_outside_universe(ou) {}
+  GeoLilyPad(PadType t, uint32_t i, uint32_t oi, Position r,
+             std::array<int32_t, 3> ti, bool ou)
+      : type(t),
+        id(i),
+        offset_index(oi),
+        r_local(r),
+        tile(ti),
+        in_lattice_outside_universe(ou) {}
 
   PadType type = PadType::Universe;
   uint32_t id = 0;
+  uint32_t offset_index = 0;
 
   // r_local is the position within the lattice,
   // NOT the position within the tile !!
