+++ src/gui_plotter.cpp
@@ -454,7 +454,7 @@ void GuiPlotter::render_image() {
           j++;
         }
       }  // while j < plot_width_
-    }  // For i which is parallel
+    }    // For i which is parallel
   }
 
   if (outline_boundaries) {
@@ -507,7 +507,7 @@ void GuiPlotter::render_image() {
           i++;
         }
       }  // while i < image.height
-    }  // For j which is parallel
+    }    // For j which is parallel
   }
 }
