+++ src/power_iterator.cpp
@@ -544,7 +544,8 @@ void PowerIterator::run() {
     // Write the final results of all estimators
     out.write("\n");
     std::stringstream output;
-    output << " Results using " << gen - nignored << " active generations:\n";
+    output << " Results using " << gen - nignored - 1
+           << " active generations:\n";
     output << " -----------------------------------\n";
     output << std::fixed << std::setprecision(6);
     output << " | kcol    = " << Tallies::instance().kcol_avg() << " +/- "
